% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terraclim.R
\name{get_terraclim}
\alias{get_terraclim}
\title{download TerraClimate data}
\usage{
get_terraclim(
  df,
  id_var,
  lat_var,
  lon_var,
  clim_vars = c("ALL"),
  year = NULL,
  show_prog = TRUE,
  conserve = FALSE
)
}
\arguments{
\item{df}{An object of class data.frame or matrix containing a column with observation
identifiers and two columns with corresponding coordinates, i.e., longitude and latitude.}

\item{id_var}{A character of length one or the unquoted name of the column storing observation
identifiers, e.g. "sample_id". Identifiers can be missing (\code{NA}). They will be set
to \emph{\if{html}{\out{<NA>}}\if{html}{\out{<dot>}}\if{html}{\out{<row number>}}}.}

\item{lat_var}{A character of length one or the unquoted name of the column storing latitude
values, e.g. "LAT".}

\item{lon_var}{A character of length one or the unquoted name of the column storing latitude
values, e.g. "Longitude".}

\item{clim_vars}{A character vector specifying the terraclimate variables to download.
Any combination of "aet", "def", "pet", "ppt", "q", "soil", "srad", "swe", "tmax", "tmin",
"vap", "ws", "vpd", "PDSI", or "ALL" to load all of them. See detail section for more information
on the variables. Default is "ALL".}

\item{year}{A numeric value specifying the year for which data should be downloaded, e.g. 1984.}

\item{show_prog}{A logical value indicating whether to print progress bar. Default is TRUE.
Accuracy depends on the stability of the transmission rate.}

\item{conserve}{A logical value indicating whether to include columns other than \code{id_var} of the
original data frame provided in df. Default is FALSE. See details for recommendations when to
use.}
}
\value{
tidy data.frame
}
\description{
Download data from the TerraClimate data set based on provided coordinates. One or more climatic
variables can be specified. Data can be retrieved for all available years or, optionally, for a
specific year. Climatic data are returned in long format which ensures seamless integration
into a \emph{tidy} workflow.
}
\details{
Coordinates are not checked for validity, i.e., if they are located over bodies of land and fall
in the range of -180:180 for longitude and -90:90 for latitude.

The TerraClimate data set is extended on a yearly basis, which lead to a varying length of the
returned data.frame. The expected length can be calculated by: \
\eqn{(year_{current} - 1958) * 12 * length_{clim\_var} * nrow_{df}}

When working with larger data sets it is advisable to set conserve = FALSE to reduce memory
footprint. Retrieving all 14 TerraClimate variables (in 2023) for a data.frame with 200
observations results in a data.frame with 2,217,600 rows.

Available TerraClimate variables:\tabular{llll}{
   Code \tab Variable \tab Detail \tab Unit \cr
   \strong{aet} \tab Actual Evapotranspiration \tab monthly total \tab \eqn{mm} \cr
   \strong{def} \tab Climate Water Deficit \tab monthly total \tab \eqn{mm} \cr
   \strong{pet} \tab Potential evapotranspiration \tab monthly total \tab \eqn{mm} \cr
   \strong{ppt} \tab Precipitation \tab monthly total \tab \eqn{mm} \cr
   \strong{q} \tab Runoff \tab monthly total \tab \eqn{mm} \cr
   \strong{soil} \tab Soil Moisture \tab total column at end of month \tab \eqn{mm} \cr
   \strong{srad} \tab Downward surface shortwave radiation \tab  \tab \eqn{W/m^2} \cr
   \strong{swe} \tab Snow water equivalent \tab at end of month \tab \eqn{mm} \cr
   \strong{tmax} \tab Max Temperature \tab average for month \tab \eqn{°C}  \cr
   \strong{tmin} \tab Min Temperature \tab average for month \tab \eqn{°C}  \cr
   \strong{vap} \tab Vapor pressure \tab average for month \tab \eqn{kPa} \cr
   \strong{ws} \tab Wind speed \tab average for month \tab \eqn{m/s} \cr
   \strong{vpd} \tab Vapor Pressure Deficit \tab average for month \tab \eqn{kPa} \cr
   \strong{PDSI} \tab Palmer Drought Severity Index \tab at end of month \tab unitless \cr
}
}
\examples{


df <- data.frame(id = c("Vilcabamba", "Paracas", "Puerto Viejo", "Trinidad", "Cartagena"),
                lat = c(-4.260641, -13.833881, 9.656562, 21.808230, 10.172151),
                lon = c(-79.22274, -76.25046, -82.75600, -79.98102, -75.74658))

x <- get_terraclim(df = df,
                  id_var = id,
                  lon_var = lon,
                  lat_var = lat,
                  clim_vars = c("tmin", "ppt"),
                  show_proc = TRUE,
                  conserve = TRUE)

str(x)


# calculate expected length
clim_vars <- c("tmin", "ppt")
(as.numeric(format(Sys.Date(), "\%Y")) - 1958) * 12 * length(clim_vars) * nrow(df)

}
