% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terraclim_single.R
\name{get_terraclim_single}
\alias{get_terraclim_single}
\title{Extract climate data from a NetCDF file for a single coordinate pair}
\usage{
get_terraclim_single(
  lat,
  lon,
  clim_var,
  nc = NULL,
  lon_nc = NULL,
  lat_nc = NULL,
  year = NULL
)
}
\arguments{
\item{lat}{numeric. The latitude value.}

\item{lon}{numeric. The longitude value.}

\item{clim_var}{character. One of 'aet', 'def', 'pet',
'ppt', 'q', 'soil', 'srad', 'swe', 'tmax', 'tmin', 'vap', 'ws', 'vpd', 'PDSI'}

\item{nc}{A NetCDF file including the \emph{clim_var} variable.
Will be accessed online if none provided.}
}
\value{
A vector of length (<year of latest records> - 1958 ) * 12. Values start at
January 1958.
}
\description{
Load terraclimate data of specified climatic variable and extract information
of the closest data point(s) to the specified coordinates.
}
\details{
TODO: specify nc param - what class?

If a coordinate has two or more equally distant data points in the
 NetCDF file, the mean value of these 'nearest neighbors' is calculated.
}
\examples{
# ambiguous match for longitude
get_terraclim_single(12, 22, "def")
# no ambiguous match
get_terraclim_single(12.12345678, 22.12345678, "def")

}
