% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_progress.R
\name{show_progress}
\alias{show_progress}
\title{Show Loop Progress}
\usage{
show_progress(i, n, names = NA_character_, time = FALSE)
}
\arguments{
\item{i}{integer. The loop counter. Expected to start at 1.}

\item{n}{integer. The maximum loop count, e.g. length of the data.}

\item{names}{A character vector of length n. The names of the dimension
being looped over, e.g. row.names. Defaults to iteration number.}

\item{time}{logical. Indication of whether to print an estimate of time to
complete remaining iterations. See Details section for implementation.}
}
\description{
Print a progress bar, percentage completed, current iteration, and, optionally, estimation
of remaining time until completion of loop to console.
}
\details{
The time estimation is implemented by calculating a time difference
of the current and the last k iterations where k is i-1. Values are stored
in a cache for estimation of remaining time. The cache is cleared when the
loop is finished, i.e. when \eqn{i == n}, when the the loop is not finished
(i.e. \eqn{i == n} was not reached) but \eqn{i == 1}, or when the
function is called in a different environment. However, the cache is not
cleared if the last call did not complete the loop and the current call
does not start at \eqn{i == 1}. This may be the case if the loop is
continued, but a falsely set \eqn{i = x | x > 1} as start value will not
be recognized as an error.
}
\examples{
for (i in 1:10) {
  TerraclimateR:::show_progress(i, 10, time = TRUE)
  Sys.sleep(2)
}

}
